test(subagents): consolidate list metadata regression
Keep the metadata allocation regression beside the existing real-SQLite list
coverage, sharing its workspace and canonical session writer. Preserve visible
model, token and status assertions before checking unrelated prompt decoding.

The existing list projection repair avoids decoding unused saved prompts.
Secretless hosted proof confirms the intended baseline failure, zero unrelated
payload decodes after the fix, complete tool-output parity, and 37 passing tests.
